+++ examples/src/commonTest/kotlin/ktorplugin/cdn/Spaces.kt
@@ -18,6 +18,7 @@ class Spaces {
     @Test
     fun `Retrieve Current Space`() = runTest {
 
+        // implementation("com.storyblok:ktor-client-storyblok:0.5.1")
         val client = HttpClient {
             install(Storyblok(CDN)) {
                 accessToken = "ask9soUkv02QqbZgmZdeDAtt"